+++ reports/qa/latest/tmp_list_collections.js
@@ -1,2 +1,0 @@
-const dbx = db.getSiblingDB('mern_app_prod');
-printjson(dbx.getCollectionNames());